+++ .gitignore
@@ -1,0 +1,2 @@
+*.class
+test.txt
